requirements1.txt
amqp==2.5.0
aniso8601==5.1.0
autopep8==1.4.4
Babel==2.7.0
billiard==3.6.0.0
bleach==2.1.4
boto3==1.9.163
botocore==1.12.163
braintree==3.49.0
cachetools==3.1.1
cairocffi==1.0.2
CairoSVG==2.4.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cssselect==1.0.3
cssselect2==0.2.1
cssutils==1.0.2
dataclasses==0.6
defusedxml==0.6.0
dj-database-url==0.5.0
dj-email-url==0.2.0
Django==2.2.2
django-absolute==0.3
django-appconf==1.0.3
django-babel==0.6.2
django-bootstrap4==0.0.8
django-cache-url==3.0.0
django-countries==5.3.3
django-debug-toolbar==1.11
django-debug-toolbar-request-history==0.0.10
django-elasticsearch-dsl==0.5.1
django-emailit==0.2.2
django-filter==2.1.0
django-graphql-jwt==0.2.1
django-icons==0.2.1
django-impersonate==1.4.1
django-js-asset==1.2.2
django-measurement==3.0.0
django-mptt==0.10.0
django-phonenumber-field==2.4.0
django-prices==1.0.4
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-recaptcha==1.5.0
django-redis==4.10.0
django-render-block==0.6
django-silk==2.0.0
django-storages==1.7.1
django-templated-email==2.3.0
django-versatileimagefield==1.10
django-webpack-loader==0.6.0
docutils==0.14
elasticsearch==6.3.1
elasticsearch-dsl==6.0.1
Faker==1.0.7
freezegun==0.3.12
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.1
google-cloud-storage==1.16.1
google-i18n-address==2.3.4
google-measurement-protocol==1.0.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gprof2dot==2016.10.13
graphene==2.1.5
graphene-django==2.2.0
graphene-django-optimizer==0.4.0
graphql-core==2.2
graphql-relay==0.4.5
html5lib==1.0.1
idna==2.8
ipaddress==1.0.22
Jinja2==2.10.1
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.6.1
lxml==4.3.4
Markdown==3.1.1
MarkupSafe==1.1.1
maxminddb==1.4.1
maxminddb-geolite2==2018.703
measurement==2.0.1
mpmath==1.1.0
oauthlib==3.0.1
phonenumberslite==8.10.13
Pillow==5.4.1
premailer==3.4.1
prices==1.0.0
promise==2.2.1
protobuf==3.8.0
psycopg2-binary==2.8.2
purl==1.5
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
Pyphen==0.9.5
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
raven==6.9.0
razorpay==1.2.0
redis==3.2.1
relativetimebuilder==1.0.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
Rx==1.6.1
s3transfer==0.2.1
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
stripe==2.29.4
sympy==1.4
text-unidecode==1.2
tinycss2==1.0.2
typing==3.6.6
urllib3==1.25.3
vine==1.3.0
WeasyPrint==47
webencodings==0.5.1
